refatora, cria CourseController e CourseService, além de outros services e controllers deletados
